Enabled NRT and added XML docs for AL.SocketClient - enabled NRT for AL.SocketClient - added XML documentation for AL.SocketClient - added IInstancedLocation.cs to handle "In" property - added extensions for IInstancedLocation - changed UpgradeData to QueuedActionResultData - last commit I added IPolygon support and forgot to mention it - removed IGameResponse interfaces and instead added documentation to every property of GameResponseData

+++ AL.SocketClient/Model/QueuedAction.cs
@@ -2,15 +2,28 @@
 
 namespace AL.SocketClient.Model
 {
+    /// <summary>
+    ///     Represents an action that will take some amount of time.
+    /// </summary>
     public record QueuedAction
     {
+        /// <summary>
+        ///     The amount of time remaining until the action is completed, in milliseconds.
+        /// </summary>
         [JsonProperty("ms")]
-        public float CurrentMS { get; init; }
+        public int CurrentMS { get; init; }
 
+        /// <summary>
+        ///     The totel amount of time it takes for the action to complete, in milliseconds.
+        /// </summary>
         [JsonProperty("len")]
-        public float LengthMS { get; init; }
+        public int LengthMS { get; init; }
 
+        /// <summary>
+        ///     TODO: Not sure... possibly if multiple actions are being taken at once, this number denotes the order they were
+        ///     taken.
+        /// </summary>
         [JsonProperty]
-        public float Num { get; init; }
+        public int Num { get; init; }
     }
 }
